Refactor naming for not-found page
